+++ instrumentation/net/http/splunkhttp/config.go
@@ -24,48 +24,19 @@ const (
 	envVarTraceResponseHeaderEnabled = "SPLUNK_TRACE_RESPONSE_HEADER_ENABLED" // Adds `Server-Timing` header to HTTP responses
 )
 
-// WithTraceResponseHeader enables or disables the TraceResponseHeaderMiddleware.
-//
-// The default is to enable the TraceResponseHeaderMiddleware if this option is not passed.
-// Additionally, the SPLUNK_TRACE_RESPONSE_HEADER_ENABLED environment variable
-// can be set to TRUE or FALSE to specify this option. This option value will be
-// given precedence if both it and the environment variable are set.
-func WithTraceResponseHeader(v bool) Option {
-	return optionFunc(func(cfg *config) {
-		cfg.TraceResponseHeaderEnabled = v
-	})
-}
-
-// Option is used for setting optional config properties.
-type Option interface {
-	apply(*config)
-}
-
 // config represents the available configuration options.
 type config struct {
 	TraceResponseHeaderEnabled bool
 }
 
-// optionFunc provides a convenience wrapper for simple Options
-// that can be represented as functions.
-type optionFunc func(*config)
-
-func (o optionFunc) apply(c *config) {
-	o(c)
-}
-
-// newConfig creates a new config struct and applies opts to it.
-func newConfig(opts ...Option) *config {
+// newConfig creates a new config struct.
+func newConfig() *config {
 	traceResponseHeaderEnabled := true
 	if v := os.Getenv(envVarTraceResponseHeaderEnabled); strings.EqualFold(v, "false") {
 		traceResponseHeaderEnabled = false
 	}
 
-	c := &config{
+	return &config{
 		TraceResponseHeaderEnabled: traceResponseHeaderEnabled,
 	}
-	for _, opt := range opts {
-		opt.apply(c)
-	}
-	return c
 }